Make SIZEOF_SIZE_T get defined
Part of this was adding AC_CHECK_SIZEOF(size_t).

The other part was removing the code to suppress
AC_INCLUDE_DEFAULTS.  That was a neat idea for making the configure
script run faster, but apparently it makes lots of normal autoconf
snippets stop working.  I'd rather be slower than error-prone, so
out _that_ goes.  [And we're not so slow, really. The configure
script still finishes in less than 4 seconds on my desktop.]

This fixes bug 5074, which happened because SIZEOF_SIZE_T was not
getting set, so smartlists could never get expanded.