bug5079: Scrub IPs before displaying them in logs.
Scrubbing is turned on by default and can be disabled with "--no-safe-logging".

The scrubbing implemented here goes farther than Tor, in that it scrubs the
client addresses too; this avoids putting live obfsproxy bridge addresses in
the log.  As a side effect, protocol traffic between obfsproxy and Tor is no
longer logged -- it can also contain addresses.