Add check for gcc and ld hardening flags.
This adds a new one shot automake macro AX_ENABLE_HARDENING and it attempts
to add the various security hardening flags we'd like to see if they're
available. When they are not available, we simply skip adding the option.
This should ensure that we have everything we could possibly have without
adding incompatible flags or having to add a lot of weird cruft to the
configure.ac file.