Don't enable bufferevent reads/writes before connecting.
This caused an assertion failure when we connected to a SOCKS
destination. We didn't call pending_socks_cb(), and instead went
directly to downstream_read_cb() which asserted that
'circuit_t.is_open' is true.

We now only call bufferevent_enable() on the bufferevent in
pending_socks_cb() when we finally connect to our SOCKS destination.